refactor: rename link-tooltip.ts to link-editor.ts to preemptively cleanup the next commit's diff
